Split core tests to public and private.
There was ODR violations due to compilation of some core code and
simultaneously linking in the sprealy_core library so the tests
for the public interface was separated to new test binary.
