+++ tests/biomolecules/sprelay/core/impl/serial_port_utils_test.h
@@ -23,9 +23,9 @@
 ****************************************************************************/
 
 /*!
- * \file      unified_serial_port_test.h
- * \brief     The biomolecules::sprelay::core::UnifiedSerialPortTest class which implements tests for
- *            biomolecules::sprelay::core::UnifiedSerialPort.
+ * \file      serial_port_utils_test.h
+ * \brief     The biomolecules::sprelay::core::SerialPortUtilsTest class which implements tests for serial port
+ *            utilities.
  *
  * \author    Jakub Klener <lumiksro@centrum.cz>
  * \date      2018-04-10
@@ -36,61 +36,30 @@
  */
 
 
-#ifndef BIOMOLECULES_SPRELAY_CORE_UNIFIED_SERIAL_PORT_TEST_H_
-#define BIOMOLECULES_SPRELAY_CORE_UNIFIED_SERIAL_PORT_TEST_H_
+#ifndef BIOMOLECULES_SPRELAY_CORE_IMPL_SERIAL_PORT_UTILS_TEST_H_
+#define BIOMOLECULES_SPRELAY_CORE_IMPL_SERIAL_PORT_UTILS_TEST_H_
 
 #include <QObject>
-#include <QString>
-
-#include <memory>
 
 #include "lumik/qtest_suite/qtest_suite.h"
 
 namespace biomolecules {
 namespace sprelay {
 namespace core {
+namespace serial_utils {
 
-class UnifiedSerialPort;
-
-class UnifiedSerialPortTest : public QObject
+class SerialPortUtilsTest : public QObject
 {
     Q_OBJECT
-public:  // NOLINT(whitespace/indent)
-    static const int kCommandTimeoutMs;
-    static const int kFactoryDefaultsTimeoutMs;
-    static const int kDelayBetweenCommandsMs;
-
 private slots:  // NOLINT(whitespace/indent)
-    // TODO(lumik): store relay states before tests (timers, button modes, relay statuses) and reset them at the end
-    void initTestCase();
-    void availablePorts();
-    void switchRealVirtual();
-    void realBenchmark_data();
-    void realBenchmark();
-    void realJumperStatus();
-    void realFirmwareVersion();
-    void realQueryAllTimers();
-    void realSetMoreTimers();
-    void realTimer();
-    void realDefaultTimer();
-    void realMoreTimers();
-    void realMoreDefaultTimers();
-    void cleanupTestCase();
-
-private:  // NOLINT(whitespace/indent)
-    std::unique_ptr<UnifiedSerialPort> createSerialPort(QString port_name) const;
-    void resetRelays(UnifiedSerialPort *serial_port) const;
-    bool compareResponse(const unsigned char *response, const unsigned char *expected);
-    void sendCommand(UnifiedSerialPort *serial_port, const unsigned char *command) const;
-    bool measureCommandWithResponse(UnifiedSerialPort *serial_port, const unsigned char *message, qint64 *elapsed_ms);
-    bool real_card_present_;
-    QString real_card_port_name_;
+    void hexToByteTestCase();
 };
 
-ADD_TEST(UnifiedSerialPortTest)
+ADD_TEST(SerialPortUtilsTest)
 
+}  // namespace serial_utils
 }  // namespace core
 }  // namespace sprelay
 }  // namespace biomolecules
 
-#endif  // BIOMOLECULES_SPRELAY_CORE_UNIFIED_SERIAL_PORT_TEST_H_
+#endif  // BIOMOLECULES_SPRELAY_CORE_IMPL_SERIAL_PORT_UTILS_TEST_H_
